Customer sends his location to the chat group
Customer can send his location to the chat server. And the chat server
and send the customer's location to all customers via a google map url.

In index.html, a button is added for the customer to click to send out
his locatioin.
In index.js, a click event and its handler is created for this button.
The handler will read the location and send the data with an event to
the server.
In server.js, a click event handler is created to send the location
message to all customers.
In index.js, a event handler is created to process the location message
and display it on the web GUI.

A mocha test code is created to test the locationMessage.