Support User login, logout, and update/show user lists
Support a user to log into the chat room. The user name and room name
will be checked and then the user is added into the list.

Support to display the user list in the "People" sidebar in the chat web
page of the user.

Users in different chat room cannot see each other.

Support to remove the user when the user logs out from the chat room.